Mira clicd process

+++ app/workers/import_mira_files_process.rb
@@ -3,19 +3,19 @@ class ImportMiraFilesProcess
   include Sidekiq::Worker
   sidekiq_options queue: :import_mira_files_process, retry: 3
 
-  def perform(representative_number, representative_abbreviated_name)
-    Mira.by_representative(representative_number).delete_all
-    WeeklyMira.by_representative(representative_number).delete_all
-
-    import_file("https://s3.amazonaws.com/piarm/#{representative_abbreviated_name}/MIRA2FILE", 'miras')
-    import_file("https://s3.amazonaws.com/piarm/#{representative_abbreviated_name}/MIRA2FILW", 'weekly_miras')
-
-    Mira.by_representative(representative_number).by_record_type.each_with_progress do |mira|
-      ImportMiraData.perform_async(mira.attributes)
-    end
-
-    WeeklyMira.by_representative(representative_number).by_record_type.each_with_progress do |mira|
-      ImportWeeklyMiraData.perform_async(mira.attributes)
+  def perform(representative_number, representative_abbreviated_name, weekly = false)
+    if weekly
+      WeeklyMira.by_representative(representative_number).delete_all if weekly
+      import_file("https://s3.amazonaws.com/piarm/#{representative_abbreviated_name}/MIRA2FILW", 'weekly_miras')
+      WeeklyMira.by_representative(representative_number).by_record_type.each_with_progress do |mira|
+        ImportWeeklyMiraData.perform_async(mira.attributes)
+      end
+    else
+      Mira.by_representative(representative_number).delete_all
+      import_file("https://s3.amazonaws.com/piarm/#{representative_abbreviated_name}/MIRA2FILE", 'miras')
+      Mira.by_representative(representative_number).by_record_type.each_with_progress do |mira|
+        ImportMiraData.perform_async(mira.attributes)
+      end
     end
   end
